Prevent unpaired bold tokens in .BI macros affecting next line
The word "Plain" in the sequence below would be mistakenly italicised:

    .BI Bold Italic Bold
    Plain

The highlighting should terminate after the first line, since there's no
trailing "\" instructing nroff to extend the formatting to the next line
(this is formally-known as a "concealed newline" in roff terminology).

The matching isn't perfect; there're still bugs with how backslashes are
handled. If a concealed newline immediately follows the closing token of
a bold string, the first word of the next line is incorrectly treated as
a different word. The correct behaviour, going by GNU nroff's output, is
to conjoin them:

    .BI Bo\       ## Expected:  *Bold* /Italic/ *Bold* /Italic/
    ld Italic     ## Actual:    *Bo*/ld/ *Italic*