Apply comment-highlighting to blank control lines
In roff, lines consisting solely of a full-stop are skipped, which makes
them useful to increase readability. Many roff files use combinations of
full-stops and .\" to markup comment blocks, so it makes sense to colour
both lines the same.

This is technically a hack being added for purely aesthetic purposes: an
actual empty line would insert a line-break in the document's output, so
authors abuse the empty control-lines instead.