Add recognition for single-quotes as control characters
Control lines may start with either a full-stop or a singular quote:

    .TH TITLE     \" Both of these lines
    'TH TITLE     \" are equivalent

The difference between the two is that single-quotes stop a "break" from
occurring after the request's invocation. A "break" in troff terminology
refers to the buffer's currently-collected contents being flushed to the
output stream, causing a newline to occur.

Only a handful of commands trigger a break, so usage of the single-quote
is generally limited to very specific use-cases.

Sources:

* "Troff User's Manual" [CSTR #54], §1.1
* "A TROFF Tutorial", p.8