Improve handling of whitespace around request names
Horizontal whitespace is permitted in between a control line's character
and the name of a command. Several requests added earlier in development
didn't accommodate this, so lines like ".  TH" weren't highlighted.

A shortcoming was discovered in the solution applied by eb5fb91: because
the Oniguruma engine doesn't support variable-length look-behinds, there
isn't a way to check for possible whitespace in preceding tokens:

    (?<=[.'][ \t]*BR)  # Error
    (?<=[.'][ \t]BR)   # Acceptable

The problem this code was written to address is already an edge-case: it
isn't likely that an author would escape a line-break inside a request's
quoted second argument. It's arguably less likely the author would do so
with whitespace injected between the control character and control name.

Consequently, we could probably get away with a "close enough" solution.
Instead of testing for any length of intermediate whitespace, we perform
five different checks in the lookbehind. Each check tests for a sequence
of at least four whitespace characters (which may be tabs, spaces, or an
assortment of both):

    .BI "AA" "B\       # No space
    . BI "AA" "B\      # 1-space
    .  BI "AA" "B\     # 2-spaces
    .   BI "AA" "B\    # 3-spaces
    .    BI "AA" "B\   # 4-spaces

Any line using more than five whitespace characters to separate the name
of a control from its character will incorrectly highlight the following
tokens of an escaped second argument.

Note the look-behind isn't simply checking the last two characters: it's
possible that would be incorrectly triggered by arguments whose last two
characters match the macro's name.