Fix assertions in macro/request look-aheads
In roff, \*n may be used to expand a defined string into an argument. An
invocation like the following statement is a legal invocation:

    .ds t " BOLD TEXT"
    .B\*t

The previous lookaheads only tested for the existence of whitespace (and
some weren't even taking tabs into consideration, either). Since escapes
may be introduced anywhere, the lookaheads have been extended to include
backslashes.

The error highlighting for underline macros has been removed. Expansions
like the above would be incorrectly marked as invalid; furthermore, it's
not the role of a language package to validate a user's input.