Fix capitalisation of subsection macro
This error is the fault of an online reference with incorrect info:

    http://www.lemoda.net/unix/troff-dictionary/index.html

The page erroneously lists several macros with inaccurate capitalisation
and describes some as GNU-specific. Since this source is unreliable, the
mention of GNU has been pulled from the TH macro's scope, as it may very
well be supported by other troff processors too.