Skip extra backslashes when matching escapes
It's common for escape sequences to be prepended with extra backslashes,
particularly in macro or string definitions. This needs to be taken into
account when hinting conditionals: lines like ".if \\nX" weren't matched
because only ".if \nX" was expected.

Additionally, casting the extra slashes as punctuation boosts legibility
in certain themes which fade such symbols for the reader (like DuoTone).