Add recognition of predefined registers
Readonly registers are now highlighted as invalid in assignment, and any
expansion sequences now mark predefined register names as constants. GNU
extensions are also distinguished from traditional roff registers by the
inclusion of "gnu" in the scope's name.

Note: The classical troff documentation lists $$ as a readonly register,
whereas only names that start with a full-stop are readonly in GNU roff.
This means it's possible to modify the process ID in a Groff session. As
this is unlikely, $$ is still listed as a readonly register to adhere to
the original language implementation.