Add pattern-matching for monospace manuscript text
Note the .CW macro, like .UL, can highlight multiple lines by passing no
arguments. This behaviour is deliberately excluded from this package, as
the risk of incorrect highlighting is too great. Authors may define a CW
or UL macro of their own, possibly as a wrapper around existing macros.

The likelihood of this happening is far greater than the likelihood of a
user actually using the manuscript package's .CW/.UL macros in practice.