Fix comments being highlighted like arguments
We need to split #params into two different blocks, as it's used for two
different applications. One is to add the "function-call" scope, and the
other is to simply include multiple related patterns at once. The change
to the end-pattern made in dfabee0 caused trailing comments to be picked
up as arguments at the end of a request call.