Add missing "meta.function" scope to macro calls
We avoid adding this to the top-level match, because it's too easy for a
marked-up region of text to be highlighted like a function. E.g., themes
might target anything with a class of ".function" in blue, which makes a
block of formatted text harder to distinguish from control lines.

It's a very minor detail, but it might make a difference.