Extend matching of comment-banners to include other forms
DWB included many troff files which lacked conventional file extensions.
The only consistency was the use of non-breaking requests affixed to an
indented comment:

    '''  \" Copyright c 1990 AT&T
    '''  \"  All Rights Reserved

Consequently, the grammar is of little use when studying their codebase.
Extending the pattern-match to include .\" as well results in two-thirds
of DWB's unmarked Roff files triggering the grammar when opened.