Truncate refer(1) regions inside macro definitions
For reasons similar to those detailed in 5030073, macro definitions with
lines starting with `.[` should terminate any refer(1) highlighting that
doesn't have an accompanying `.]` line. For example:

    .de ][
    .   [\\$1
    # Comment
    Keyword
    ..
    # Not a comment

We don't extend this to other preprocessors, since their syntax has less
risk of being used as part of a macro definition (and we can't afford to
blindly terminate at any line that happens to begin with `..`).