Link to more actively-maintained man page sources
As pointed out by @g-branden-robinson in recent correspondence, the docs
hosted on linux.die.net are notoriously out-of-date, whereas man7.org is
frequently updated.

Similarly, use more appropriate sources of documentation for mdoc(7) and
groff_me(7). The DFORMAT and chem(1) links are broken (and only point to
downloadable PostScript files in the Wayback Machine), so leave them for
now.